+++ src/Clusterers/KMeans.php
@@ -30,10 +30,10 @@ use const Rubix\ML\EPSILON;
  *
  * A fast online centroid-based hard clustering algorithm capable of clustering
  * linearly separable data points given some prior knowledge of the target number
- * of clusters (defined by *k*). K Means with inertia is trained using adaptive
- * mini batch gradient descent and minimizes the inertia cost function. Inertia
- * is defined as the sum of the distances between each sample and its nearest
- * cluster center (centroid).
+ * of clusters (defined by *k*). K Means is trained using adaptive mini batch
+ * Gradient Descent and minimizes the inertia cost function. Inertia is defined
+ * as the average sum of distances between each sample and its nearest cluster
+ * centroid.
  *
  * References:
  * [1] D. Sculley. (2010). Web-Scale K-Means Clustering.