+++ CHANGELOG.md
@@ -5,6 +5,8 @@
     - Added predict single sample method to Learner interface
     - Changed method signature of random subset without replacement
     - Changed K Means default max iterations
+    - Cross validators now stratify dataset automatically
+    - Changed default k in K Fold validator
 
 - 0.0.14-beta
     - Added feature importances to Gradient Boost